pyproject.toml
[project]
name = "acv-assignment1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "torch>=2.2.0,<3.0.0",
    "torchvision>=0.17.0,<1.0.0",
    "tqdm>=4.66.0",
]
